Fix code quality issues and reduce duplication
- Extract testNotReadyProbe helper in health_test.go to eliminate
  duplicate test code for not-ready probe assertions
- Inline category/tag page queries in frontend.go, removing the
  pageFetcher abstraction that caused duplicate function patterns
- Add registerFrontendRoutes and registerSettingsRoutes helpers
  in main.go to consolidate repeated route registration
- Extract getProviderContext in embed module handlers to share
  common request validation logic
- Fix float64 literal in render_test.go (remove redundant type)

+++ internal/render/render_test.go
@@ -347,7 +347,7 @@ func TestTemplateFuncs_Int64(t *testing.T) {
 		{"int", 42, 42},
 		{"int32", int32(42), 42},
 		{"int64", int64(42), 42},
-		{"float64", float64(42.9), 42},
+		{"float64", 42.9, 42},
 		{"string", "42", 42},
 		{"invalid string", "abc", 0},
 		{"nil", nil, 0},
